Removal of un-used folders and files
Removed old folder structure along with files. Updated page location. Cleaned up readme. Moved data example to own folder/component. Added context example. Cleaned up index.js and removed old imports.
