+++ src/components/start-layout/routes.js
@@ -2,9 +2,9 @@ import React from 'react'
 import {Route, Switch} from "react-router-dom"
 
 import NotFoundPage from "../not-found-page"
-import IndexPage from "./index-page"
-import AboutPage from "./about-page"
-import DataPage from "./data-page";
+import IndexPage from "../pages/index-page"
+import AboutPage from "../pages/about-page"
+import DataPage from "../pages/data-page";
 
 const Routes = () => (
   <Switch>
