+++ src/index.js
@@ -2,13 +2,8 @@ import React from 'react'
 import {render} from 'react-dom'
 import {Router} from "react-router-dom"
 import {createBrowserHistory} from "history"
-// import Routes from "./components/start-simple/routes"
 import Routes from "./components/start-layout/routes"
-// import Routes from "./components/start-people/routes"
-// import Routes from "./components/start-context/routes"
-//import Routes from "./components/start-forms/routes"
-
-import AppProvider from "./components/start-context/app-provider"
+import AppProvider from "./components/context/app-provider"
 
 import './index.css';
